py_backend/requirements.txt
requests
Pillow
tensorflow
pyyaml